requirements-dev.txt
Django>=3.1,<3.2
requests>=2.21,<3
PyGithub>=1.51,<2.0
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
mysqlclient
python-gitlab>=2.0,<3.0
python-slugify>= 4.0.0,<5
cryptography>=2.7,<3
PyJWT>=1.7,<2
bokeh==2.2.1
python-dateutil>=2.8.1,<3
pandas>=1.0.5,<1.1
django-model-utils>=4.0